requirements.txt
beautifulsoup4==4.6.3
certifi==2018.10.15
chardet==3.0.4
idna==2.7
isodate==0.6.0
lxml==4.2.5
pyparsing==2.2.2
python-dateutil==2.7.5
rdflib==4.2.2
requests==2.20.1
requests-futures==0.9.8
six==1.11.0
urllib3==1.24.1
